+++ src/App.js
@@ -4,17 +4,25 @@ import "./App.css";
 import Auth from "./components/Auth";
 import Layout from "./components/Layout";
 import Notification from "./components/Notification";
-import { sendCartData } from "./store/cart-slice";
-import { uiActions } from "./store/ui-slice";
+import { sendCartData, fetchData } from "./store/cart-actions";
 
 function App() {
 
   const [isFirstRender, setFirstRender] = useState(true);
+  // let isFirstRender = true;
 
   const dispatch = useDispatch();
   const notification = useSelector(state => state.ui.notification);
   const cart = useSelector(state => state.cart);
   const isLoggedIn = useSelector(state => state.auth.isLoggedIn);
+
+  useEffect(() => {
+    if (isFirstRender) {
+      setFirstRender(false);
+      return;
+    }
+    dispatch(fetchData());
+  }, [cart, dispatch, isFirstRender]);
 
   useEffect(() => {
     if (isFirstRender) {
