+++ src/components/Layout.js
@@ -32,3 +32,8 @@ const Layout = () => {
 };
 
 export default Layout;
+
+
+// Warning: Maximum update depth exceeded. This can happen when a component calls setState inside useEffect, but useEffect either doesn't have a dependency array, or one of the dependencies changes on every render.
+//     at App (http://localhost:3000/static/js/bundle.js:39:90)
+//     at Provider (http://localhost:3000/static/js/bundle.js:47313:5)
