+++ src/store/cart-slice.js
@@ -1,15 +1,21 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { uiActions } from "./ui-slice";
 
 const cartSlice = createSlice({
     name: "cart",
     initialState: {
         itemsList: [],
         totalItems: 0,
-        showCart: false
+        showCart: false,
+        changed: false
     },
     reducers: {
+        replaceData(state, action) {
+            state.totalItems = action.payload.totalItems;
+            state.itemsList = action.payload.itemsList;
+        },
+
         addToCart(state, action) {
+            state.changed = true;
             const newItem = action.payload;
             const existingItem = state.itemsList.find(item => item.id === newItem.id);
             if (existingItem) {
@@ -28,6 +34,7 @@ const cartSlice = createSlice({
         },
 
         removeFromCart(state, action) {
+            state.changed = true;
             const id = action.payload;
             const existingItem = state.itemsList.find(item => id === item.id);
 
@@ -45,45 +52,6 @@ const cartSlice = createSlice({
         }
     }
 });
-
-export const sendCartData = (cart) => {
-    return async (dispatch) => {
-        const sendRequest = async () => {
-            //Send state as sending request
-            dispatch(uiActions.showNotification({
-                message: "Sending Request",
-                type: "warning",
-                open: true
-            }));
-            
-            const res = await fetch("https://redux-http-17d55-default-rtdb.firebaseio.com/cartItems.json",
-                {
-                    method: "PUT",
-                    body: JSON.stringify(cart)
-                }
-            );
-            const data = await res.json();
-      
-      
-            //Send state as request is successful
-            dispatch(uiActions.showNotification({
-                open: true,
-                message: "Successfully sent request to the Database",
-                type: "success"
-            }));
-        };
-
-        try {
-            await sendRequest();
-        } catch (err) {
-            dispatch(uiActions.showNotification({
-                open: true,
-                message: "Sending request failed",
-                type: "error"
-            }));
-        }
-    }
-}
 
 export const cartActions = cartSlice.actions;
 export default cartSlice;
